requirements.txt
kafka-python==2.0.2
pymongo==3.12.0
backoff==1.11.1